feat(login) reuse sync table for task preferences
